[cmyth] Retrieve radio flag within channels
In order to determine if a channel is a radio (audio only) channel,
we can use the is_audio_service flag, that MythTV sets when scanning for
channels. Unfortunately this information is not directly available in
the channel table.

In the current code, it is necessary to trigger one SQL query for every channel
in order to get the is_audio_service flag (cmyth_mysql_isradio).

This patch removes the function cmyth_mysql_isradio and queries the
is_audio_service flag directly in cmyth_mysql_get_chanlist. The flag is
stored in the field radio within cmyth_channel and can be accessed by
cmyth_channel_radio.

This reduces drastically the number of SQL queries and improves
performance especially for users with a huge amount of channels.